feat: add buyWiddow  class
